# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Obiara API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Authoritative REST contract for Obiara member clients. Durable API changes
    are additive within v1. Retriable writes require an Idempotency-Key.
...
servers:
  - url: /
    description: API root
...
paths:
...
  /v1/admin/care/cases:
    get:
      operationId: listAdminCareCases
      tags: [Admin, Safety]
      summary: List open care cases
      description: Requires safety-review scope. Subjects are returned only as privacy-keyed references; the care queue is separate from enforcement.
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Oldest-first open and engaged care cases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCareCaseListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
        "503":
          $ref: "#/components/responses/InternalError"
  /v1/admin/care/cases/{id}/engagement:
    post:
      operationId: engageAdminCareCase
      tags: [Admin, Safety]
      summary: Mark a care case as engaged by trained staff
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Care case engaged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCareCaseEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
        "409":
          description: Care case is no longer open.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/InternalError"
  /v1/admin/care/cases/{id}/resolution:
    post:
      operationId: resolveAdminCareCase
      tags: [Admin, Safety]
      summary: Resolve an engaged care case with approved resource keys
      description: Requires safety-review scope and a fresh MFA claim. This records resources used but does not send messages, diagnose the member or trigger enforcement.
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCareResolutionInput"
      responses:
        "200":
          description: Care case resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCareCaseEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AdminRoleRequired"
        "409":
          description: Care case is not engaged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/InternalError"
...
  /v1/circles/{circleId}/ampe:
    post:
      operationId: createAmpeRound
      tags: [Community]
      summary: Start a private Ampe round in an exact two-member circle
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "201":
          description: Private round created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmpeEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AmpeNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/ampe/{roundId}:
    get:
      operationId: getAmpeRound
      tags: [Community]
      summary: Read a privacy-safe Ampe round projection
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AmpeRoundId"
      responses:
        "200":
          description: Current member-relative round state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmpeEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AmpeNotAvailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/ampe/{roundId}/commands:
    post:
      operationId: commandAmpeRound
      tags: [Community]
      summary: Mark ready or lock a manual private gesture
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AmpeRoundId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmpeCommandInput"
      responses:
        "200":
          description: Command accepted and current projection returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmpeEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AmpeNotAvailable"
        "409":
          $ref: "#/components/responses/AmpeConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: Short-lived member access token; the session subject must match memberId.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff token carrying the explicitly documented capability and recent-MFA claim where required.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Safe caller-provided identifier; invalid values are replaced.
      schema:
        $ref: "#/components/schemas/CorrelationId"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key reused for retries of the same command.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    CircleRoomId:
      name: circleId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    AmpeRoundId:
      name: roundId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  headers:
    CorrelationId:
      required: true
      description: Effective request correlation identifier.
      schema:
        $ref: "#/components/schemas/CorrelationId"
  responses:
    Unauthorized:
      description: A valid authenticated session is required.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AmpeNotAvailable:
      description: The round or its exact two-member circle context is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AmpeConflict:
      description: The round sequence changed before the command was applied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidJSON:
      description: The body is malformed, oversized, contains unknown fields, or contains multiple values.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidJSON:
              value:
                error:
                  code: invalid_json
                  message: The request body must be one valid JSON object.
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    UnsupportedMediaType:
      description: Content-Type is not application/json.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AdminRoleRequired:
      description: The actor lacks the required admin role.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Request fields or the idempotency key are invalid.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidEmail:
              value:
                error:
                  code: validation_failed
                  message: One or more fields are invalid.
                  details:
                    - field: email
                      reason: must be a valid email address
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    InternalError:
      description: An unexpected server failure occurred.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    CorrelationId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    Metadata:
      type: object
      additionalProperties: false
      required: [correlationId]
      properties:
        correlationId:
          $ref: "#/components/schemas/CorrelationId"
...
    AdminCareCaseData:
      type: object
      additionalProperties: false
      required:
        [caseId, subjectRef, signal, status, scripts, createdAt, version]
      properties:
        caseId:
          type: string
        subjectRef:
          type: string
          description: Stable privacy-keyed subject reference; never the member account identifier.
        signal:
          type: string
          enum:
            [
              distress_report,
              self_harm_indication,
              victim_report,
              okyeame_escalation,
              closure,
            ]
        status:
          type: string
          enum: [open, engaged, resolved]
        scripts:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              [
                helpline_directory_gh,
                counselor_referral,
                support_content,
                closure_quietening,
              ]
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 1
    AdminCareCaseEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AdminCareCaseData"
        meta:
          $ref: "#/components/schemas/Metadata"
    AdminCareCaseListData:
      type: object
      additionalProperties: false
      required: [cases]
      properties:
        cases:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/AdminCareCaseData"
    AdminCareCaseListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AdminCareCaseListData"
        meta:
          $ref: "#/components/schemas/Metadata"
    AdminCareResolutionInput:
      type: object
      additionalProperties: false
      required: [scripts]
      properties:
        scripts:
          type: array
          minItems: 1
          maxItems: 4
          uniqueItems: true
          items:
            type: string
            enum:
              [
                helpline_directory_gh,
                counselor_referral,
                support_content,
                closure_quietening,
              ]
...
    AmpeCommandInput:
      type: object
      additionalProperties: false
      required: [action, expectedSequence]
      properties:
        action:
          type: string
          enum: [ready, lock]
        choice:
          type: string
          enum: [together, apart]
        expectedSequence:
          type: integer
          format: int64
          minimum: 0
    AmpePlayerData:
      type: object
      additionalProperties: false
      required: [ready, connected, locked]
      properties:
        ready:
          type: boolean
        connected:
          type: boolean
        locked:
          type: boolean
    AmpeData:
      type: object
      additionalProperties: false
      required: [id, sequence, you, other, paused, complete]
      properties:
        id:
          type: string
        sequence:
          type: integer
          format: int64
          minimum: 0
        you:
          $ref: "#/components/schemas/AmpePlayerData"
        other:
          $ref: "#/components/schemas/AmpePlayerData"
        paused:
          type: boolean
        ownChoice:
          type: string
          enum: [together, apart]
        yourReveal:
          type: string
          enum: [together, apart]
        otherReveal:
          type: string
          enum: [together, apart]
        complete:
          type: boolean
    AmpeEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AmpeData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/Metadata"
...